Add case for empty data
While this is never possible with any inputs given from advent of
code, it feels wrong to just assume a filter wont remove all the items.
(Or someone putting in an empty list of data, thats also a case that
wouldnt be handled without this edgecase)

There is also the possibility that the reduce will get stuck on 2 or
more items in an endless loop. But i dont want to think about this
possibility when our data is nice and will always work :D